+++ src/main/java/com/elumini/miniautorizador/service/TransacaoService.java
@@ -3,5 +3,5 @@ package com.elumini.miniautorizador.service;
 import com.elumini.miniautorizador.model.Cartao;
 
 public interface TransacaoService {
-	Cartao atualizarSaldo(Cartao cartao, Long saldo);
+	Cartao atualizarSaldo(Cartao cartao, Double saldo);
 }
